third_party/WebKit/LayoutTests/fast/dom/processing-instruction-inherit-characterdata-expected.txt

Test that ProcessingInstruction inherits from CharacterData

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ProcessingInstruction.__proto__ is CharacterData
PASS ProcessingInstruction.prototype.__proto__ is CharacterData.prototype
PASS processingInstruction.__proto__.__proto__ is CharacterData.prototype
PASS processingInstruction.target is "target"
PASS processingInstruction.data is "data"
PASS processingInstruction.length is 4
PASS processingInstruction.data is ""
PASS processingInstruction.length is 0
PASS successfullyParsed is true

TEST COMPLETE
